Moved enum folder out of the packets folder and into the main f1.data folder
