+++ src/main/java/f1/data/SessionInitializationResult.java
@@ -1,9 +1,9 @@
 package f1.data;
 
 import f1.data.packets.ParticipantData;
-import f1.data.packets.enums.DriverPairingsEnum;
-import f1.data.packets.enums.Formula2Enum;
-import f1.data.packets.enums.FormulaEnum;
+import f1.data.enums.DriverPairingsEnum;
+import f1.data.enums.Formula2Enum;
+import f1.data.enums.FormulaEnum;
 import f1.data.packets.session.SessionData;
 
 import java.util.HashMap;
