+++ src/main/java/f1/data/enums/BrakingAssistEnum.java
@@ -1,4 +1,4 @@
-package f1.data.packets.enums;
+package f1.data.enums;
 
 public enum BrakingAssistEnum {
