+++ src/main/java/f1/data/enums/DriverPairingsEnum.java
@@ -1,4 +1,4 @@
-package f1.data.packets.enums;
+package f1.data.enums;
 
 import f1.data.utils.constants.Constants;
 import f1.data.utils.constants.DriverConstants;
