+++ src/main/java/f1/data/enums/FormulaEnum.java
@@ -1,4 +1,4 @@
-package f1.data.packets.enums;
+package f1.data.enums;
 
 import java.util.HashMap;
 import java.util.Map;
